Feature/comments

+++ server/controllers/follow.js
@@ -1,5 +1,11 @@
 const db = require('../db');
 const Follow = db().Follow;
+const User = db().User;
+
+const userExists = async id => {
+  const user = await User.findOne({ where: { id } });
+  return user ? true : false;
+};
 
 /**
  * @name /api/follow/:followingId
@@ -9,10 +15,17 @@ exports.follow = async (req, res, next) => {
   const { followingId } = req.params;
 
   try {
+    const exists = await userExists(followingId);
+    if (!exists) {
+      throw new Error('Unexistent User');
+    }
+
     await Follow.create({ followingId, followerId });
+    await User.increment('following', { where: { id: followerId } });
+    await User.increment('followers', { where: { id: followingId } });
     return res.status(200).json({ message: 'Following' });
   } catch (err) {
-    next({ status: 400, message: 'Unexistent User' });
+    next({ status: 400, message: err.message });
   }
 };
 
@@ -24,6 +37,11 @@ exports.unfollow = async (req, res, next) => {
   const { followingId } = req.params;
 
   try {
+    const exists = await userExists(followingId);
+    if (!exists) {
+      throw new Error('Unexistent User');
+    }
+
     const unfollowing = await Follow.destroy({
       where: { followingId, followerId },
     });
@@ -32,8 +50,11 @@ exports.unfollow = async (req, res, next) => {
       throw new Error();
     }
 
+    await User.decrement('following', { where: { id: followerId } });
+    await User.decrement('followers', { where: { id: followingId } });
+
     return res.status(200).json({ message: 'Unfollowing' });
   } catch (err) {
-    next({ status: 400, message: 'Unexistent User' });
+    next({ status: 400, message: err.message });
   }
 };
